Exception handling - delete

+++ src/main/java/com/jeffersonsousa/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.jeffersonsousa.course.entities.User;
 import com.jeffersonsousa.course.repositories.UserRepository;
+import com.jeffersonsousa.course.services.exceptions.DataBaseException;
 import com.jeffersonsousa.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,8 +32,17 @@ public class UserService {
 	}
 	
 	public void delete(Long id){
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
 		repository.deleteById(id);
-	}
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		  }
+		}
+	
+	
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
